calendar_microservice/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
uvicorn>=0.23.0
fastapi>=0.100.0
pydantic>=2.0.0
httpx>=0.24.0
python-jose>=3.3.0
pytest>=7.3.1
tenacity>=8.2.2

# MCP integration dependencies
openai-agents>=0.1.0

# Async support
asyncio>=3.4.3
aiohttp>=3.8.5

# Caching and storage
redis>=4.6.0